+++ CPE200Lab1/CPE200Lab1/CalculatorEngine.cs
@@ -5,143 +5,35 @@ using System.Text;
 using System.Threading.Tasks;
 
 namespace CPE200Lab1
-{   
-    public class CalculatorEngine
+{
+    public class CalculatorEngine : BasicCalculatorEngine
     {
-        /// <summary>
-        /// Check wether imputis a number 
-        /// </summary>
-        /// <param name="str">input string</param>
-        /// <returns></returns>
-        public bool isNumber(string str)
+        protected double firstOperand;
+        protected double secondOperand;
+
+        public void setFirstOperand(string num)
         {
-            double retNum;
-            return Double.TryParse(str, out retNum);
+            firstOperand = Convert.ToDouble(num);
         }
 
-        public bool isOperator(string str)
+        public void setSecondOperand(string num)
         {
-            switch(str) {
-                case "+":
-                case "-":
-                case "X":
-                case "÷":
-                case "%":
-                    return true;
-            }
-            return false;
+            secondOperand = Convert.ToDouble(num);
         }
 
-        public bool isOperatorUnary(string str)
-        {
-            switch (str)
-            {
-                case "√":
-                case "1/x":
-                    return true;
-            }
-            return false;
-        }
-
-        public string Process(string str)
+        public string calculate(string str)
         {
             string[] parts = str.Split(' ');
-            if(!(isNumber(parts[0]) && isOperator(parts[1]) && isNumber(parts[2])))
+            if (!(isNumber(parts[0]) && isOperator(parts[1]) && isNumber(parts[2])))
             {
                 return "E";
-            } else
+            }
+            else
             {
                 return calculate(parts[1], parts[0], parts[2], 4);
             }
 
         }
-        public string unaryCalculate(string operate, string operand, int maxOutputSize = 8)
-        {
-            switch (operate)
-            {
-                case "√":
-                    {
-                        double result;
-                        string[] parts;
-                        int remainLength;
 
-                        result = Math.Sqrt(Convert.ToDouble(operand));
-                        // split between integer part and fractional part
-                        parts = result.ToString().Split('.');
-                        // if integer part length is already break max output, return error
-                        if (parts[0].Length > maxOutputSize)
-                        {
-                            return "E";
-                        }
-                        // calculate remaining space for fractional part.
-                        remainLength = maxOutputSize - parts[0].Length - 1;
-                        // trim the fractional part gracefully. =           
-                        return decimal.Parse(result.ToString("N" + remainLength)).ToString("G29");
-
-                    }
-                case "1/x":
-                    if(operand != "0")
-                    {
-                        double result;
-                        string[] parts;
-                        int remainLength;
-
-                        result = (1.0 / Convert.ToDouble(operand));
-                        // split between integer part and fractional part
-                        parts = result.ToString().Split('.');
-                        // if integer part length is already break max output, return error
-                        if (parts[0].Length > maxOutputSize)
-                        {
-                            return "E";
-                        }
-                        // calculate remaining space for fractional part.
-                        remainLength = maxOutputSize - parts[0].Length - 1;
-                        // trim the fractional part gracefully. =
-                        return decimal.Parse(result.ToString("N" + remainLength)).ToString("G29");
-
-                    }
-                    break;
-            }
-            return "E";
-        }
-
-        public string calculate(string operate, string firstOperand, string secondOperand, int maxOutputSize = 8)
-        {
-            switch (operate)
-            {
-                case "+":
-                    return (Convert.ToDouble(firstOperand) + Convert.ToDouble(secondOperand)).ToString();
-                case "-":
-                    return (Convert.ToDouble(firstOperand) - Convert.ToDouble(secondOperand)).ToString();
-                case "X":
-                    return (Convert.ToDouble(firstOperand) * Convert.ToDouble(secondOperand)).ToString();
-                case "÷":
-                    // Not allow devide be zero
-                    if (secondOperand != "0")
-                    {
-                        double result;
-                        string[] parts;
-                        int remainLength;
-
-                        result = (Convert.ToDouble(firstOperand) / Convert.ToDouble(secondOperand));
-                        // split between integer part and fractional part
-                        parts = result.ToString().Split('.');
-                        // if integer part length is already break max output, return error
-                        if (parts[0].Length > maxOutputSize)
-                        {
-                            return "E";
-                        }
-                        // calculate remaining space for fractional part.
-                        remainLength = maxOutputSize - parts[0].Length - 1;
-                        // trim the fractional part gracefully. =
-                        return decimal.Parse(result.ToString("N" + remainLength)).ToString("G29");
-                    }
-                    break;
-                case "%":
-                    //your code here
-                    return (Convert.ToDouble(firstOperand) * Convert.ToDouble(secondOperand) / 100).ToString();
-            }
-            return "E";
-        }
     }
 }
