+++ CPE200Lab1/CPE200Lab1/ExtendForm.cs
@@ -15,13 +15,11 @@ namespace CPE200Lab1
         private bool isNumberPart = false;
         private bool isContainDot = false;
         private bool isSpaceAllowed = false;
-        //private CalculatorEngine engine;
         private RPNCalculatorEngine engine;
 
         public ExtendForm()
         {
             InitializeComponent();
-            //engine = new CalculatorEngine();
             engine = new RPNCalculatorEngine();
         }
