+++ CPE200Lab1/CPE200Lab1Test/CalculatorEngineTests.cs
@@ -18,7 +18,7 @@ namespace CPE200Lab1.Tests
             string actual;
 
             CalculatorEngine engine = new CalculatorEngine();
-            actual = engine.Process(input);
+            actual = engine.calculate(input);
             Assert.AreEqual(expected, actual);
         }
         [TestMethod()]
@@ -29,7 +29,7 @@ namespace CPE200Lab1.Tests
             string actual;
 
             CalculatorEngine engine = new CalculatorEngine();
-            actual = engine.Process(input);
+            actual = engine.calculate(input);
             Assert.AreEqual(expected, actual);
         }
         [TestMethod()]
@@ -40,7 +40,7 @@ namespace CPE200Lab1.Tests
             string actual;
 
             CalculatorEngine engine = new CalculatorEngine();
-            actual = engine.Process(input);
+            actual = engine.calculate(input);
             Assert.AreEqual(expected, actual);
         }
         [TestMethod()]
@@ -51,7 +51,7 @@ namespace CPE200Lab1.Tests
             string actual;
 
             CalculatorEngine engine = new CalculatorEngine();
-            actual = engine.Process(input);
+            actual = engine.calculate(input);
             Assert.AreEqual(expected, actual);
         }
 
@@ -63,7 +63,7 @@ namespace CPE200Lab1.Tests
             string actual;
 
             CalculatorEngine engine = new CalculatorEngine();
-            actual = engine.Process(input);
+            actual = engine.calculate(input);
             Assert.AreEqual(expected, actual);
         }
 
@@ -75,7 +75,7 @@ namespace CPE200Lab1.Tests
             string actual;
 
             CalculatorEngine engine = new CalculatorEngine();
-            actual = engine.Process(input);
+            actual = engine.calculate(input);
             Assert.AreEqual(expected, actual);
         }
 
@@ -215,7 +215,7 @@ namespace CPE200Lab1.Tests
             string actual;
 
             CalculatorEngine engine = new CalculatorEngine();
-            actual = engine.unaryCalculate("√", operand);
+            actual = engine.calculate("√", operand);
             Assert.AreEqual(expected, actual);
         }
 
@@ -227,7 +227,7 @@ namespace CPE200Lab1.Tests
             string actual;
 
             CalculatorEngine engine = new CalculatorEngine();
-            actual = engine.unaryCalculate("1/x", operand);
+            actual = engine.calculate("1/x", operand);
             Assert.AreEqual(expected, actual);
         }
 
@@ -239,7 +239,7 @@ namespace CPE200Lab1.Tests
             string actual;
 
             CalculatorEngine engine = new CalculatorEngine();
-            actual = engine.unaryCalculate("1/x", operand);
+            actual = engine.calculate("1/x", operand);
             Assert.AreEqual(expected, actual);
         }
     }
